+++ src/main/java/frc/robot/Robot.java
@@ -10,7 +10,6 @@ package frc.robot;
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
-import edu.wpi.first.wpilibj.Compressor;
 import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.Filesystem;
 import edu.wpi.first.wpilibj.TimedRobot;
@@ -56,7 +55,6 @@ public class Robot extends TimedRobot {
     Listener intakeListener;
     Listener panelListener;
     ShootFormula formula;
-    Compressor compressor;
 
     SendableChooser<Boolean> chooserQuinnDrive;
 
@@ -114,7 +112,6 @@ public class Robot extends TimedRobot {
         formula = new ShootFormula();
         intakeLimit = new DigitalInput(0);
         timer = new Timer();
-        compressor = new Compressor(0);
         intakeListener =
                 new Listener(
                         null,
@@ -133,7 +130,7 @@ public class Robot extends TimedRobot {
                         new Runnable() {
                             public void run() {
                                 try {
-                                    controlPanel.flipArmUp();
+                                    pneumatics.flipArmUp();
                                 } catch (SubsystemException e) {
                                     Console.log(e.getMessage());
                                 }
@@ -142,7 +139,7 @@ public class Robot extends TimedRobot {
                         new Runnable() {
                             public void run() {
                                 try {
-                                    controlPanel.flipArmDown();
+                                    pneumatics.flipArmDown();
                                 } catch (SubsystemException e) {
                                     Console.log(e.getMessage());
                                 }
